removed universalDetector
